+++ crates/etl/src/data/mod.rs
@@ -1,13 +1,13 @@
 //! ETL value and row data model.
 
 mod cell;
-mod row;
 mod size;
+mod table_row;
 
 pub use cell::{ArrayCell, ArrayCellNonOptional, Cell, CellNonOptional};
 pub use etl_postgres::{
     numeric::PgNumeric,
     time::{DATE_FORMAT, PgTimeTz, TIME_FORMAT, TIMESTAMP_FORMAT, TIMESTAMPTZ_FORMAT_HH_MM},
 };
-pub use row::{OldTableRow, PartialTableRow, TableRow, UpdatedTableRow};
 pub use size::SizeHint;
+pub use table_row::{OldTableRow, PartialTableRow, TableRow, UpdatedTableRow};
