+++ crates/etl/src/data/table_row.rs
