+++ crates/etl/src/destination/mod.rs
@@ -2,10 +2,11 @@
 //!
 //! Provides destination traits for sending replicated data to target systems.
 //! Destinations handle both initial table synchronization data and streaming
-//! replication events. [`PipelineDestination`] provides the facade bounds used
-//! by the pipeline runtime. [`TaskSet`] is available for destinations that
-//! complete accepted writes in background tasks while keeping shutdown and
-//! panic handling explicit.
+//! replication events. Most destination authors implement [`Destination`]
+//! directly. [`PipelineDestination`] is the blanket-implemented facade used by
+//! the pipeline runtime when it needs cloneable, worker-safe destinations.
+//! [`TaskSet`] is available for destinations that complete accepted writes in
+//! background tasks while keeping shutdown and panic handling explicit.
 
 mod async_result;
 mod base;
