+++ crates/etl/src/lib.rs
@@ -4,8 +4,8 @@
 //!
 //! This crate provides a high-performance, streaming ETL (Extract, Transform,
 //! Load) system built on Postgres logical replication. It enables real-time
-//! data synchronization from Postgres databases to various destinations with
-//! configurable transformations and robust error handling.
+//! data synchronization from Postgres databases to destination systems with a
+//! typed Rust API and robust error handling.
 //!
 //! # Key Features
 //!
@@ -25,7 +25,8 @@
 //! ## Pipeline
 //! A [`pipeline::Pipeline`] represents a complete ETL workflow that connects a
 //! Postgres publication to a destination. It manages the replication stream,
-//! applies transformations, and handles failures gracefully.
+//! coordinates initial copy and streaming workers, and handles failures
+//! gracefully.
 //!
 //! ## Destinations
 //! [`destination::Destination`] trait implementations define where replicated
@@ -143,15 +144,15 @@
 //!             memory_budget_ratio: 0.2,
 //!             max_bytes: 8 * 1024 * 1024,
 //!         },
-//!         table_error_retry_delay_ms: 10000,
+//!         table_error_retry_delay_ms: 10_000,
 //!         table_error_retry_max_attempts: 5,
 //!         max_table_sync_workers: 4,
-//!         max_copy_connections_per_table: 1,
+//!         max_copy_connections_per_table: PipelineConfig::DEFAULT_MAX_COPY_CONNECTIONS_PER_TABLE,
 //!         memory_refresh_interval_ms: 100,
 //!         replication_lag_refresh_interval_ms: 10_000,
 //!         memory_backpressure: Some(MemoryBackpressureConfig::default()),
-//!         table_sync_copy: TableSyncCopyConfig::IncludeAllTables,
-//!         invalidated_slot_behavior: InvalidatedSlotBehavior::Error,
+//!         table_sync_copy: TableSyncCopyConfig::default(),
+//!         invalidated_slot_behavior: InvalidatedSlotBehavior::default(),
 //!     };
 //!
 //!     // Create and start the pipeline
