+++ crates/etl/src/postgres/stream/mod.rs
@@ -1,0 +1,7 @@
+//! Postgres streams used by replication and table synchronization.
+
+mod events;
+mod table_copy;
+
+pub(crate) use events::{EventsStream, StatusUpdateResult, StatusUpdateType};
+pub(crate) use table_copy::TableCopyStream;
