+++ crates/etl/src/runtime/concurrency/mod.rs
@@ -1,14 +1,10 @@
 //! Concurrency primitives used by ETL workers and destinations.
 
-mod batch_budget;
-mod memory_monitor;
 mod shutdown;
 mod signal;
 mod stream;
 mod task_set;
 
-pub(crate) use batch_budget::{BatchBudgetController, CachedBatchBudget};
-pub(crate) use memory_monitor::MemoryMonitor;
 pub use shutdown::ShutdownTx;
 pub(crate) use shutdown::{ShutdownResult, ShutdownRx, create_shutdown_channel};
 pub(crate) use stream::{
