+++ crates/etl/src/runtime/mod.rs
@@ -1,11 +1,15 @@
 //! Runtime task orchestration and concurrency support.
 
 mod apply;
+mod batch_budget;
 mod error_policy;
+mod memory_monitor;
 mod table_sync;
 
 pub(crate) mod concurrency;
 
 pub(crate) use apply::{ApplyWorker, ApplyWorkerHandle};
+pub(crate) use batch_budget::{BatchBudgetController, CachedBatchBudget};
 pub(crate) use error_policy::ErrorHandlingPolicy;
+pub(crate) use memory_monitor::{MemoryMonitor, MemoryMonitorSubscription};
 pub(crate) use table_sync::{TableSyncWorker, TableSyncWorkerPool, TableSyncWorkerState};
