+++ crates/etl/src/store/mod.rs
@@ -33,5 +33,5 @@ pub use state::{StateStore, TableStates};
 
 pub use crate::replication::{
     WorkerType,
-    table_state::{TableRetryPolicy, TableState, TableStateType},
+    state::{TableRetryPolicy, TableState, TableStateType},
 };
