+++ docs/src/content/docs/explanation/architecture.md
@@ -82,6 +82,7 @@ Where replicated data goes. Implement the `Destination` trait to send data anywh
 pub trait Destination {
     fn name() -> &'static str;
     fn shutdown(&self) -> impl Future<Output = EtlResult<()>> + Send { async { Ok(()) } }
+    fn startup(&self) -> impl Future<Output = EtlResult<()>> + Send { async { Ok(()) } }
     fn drop_table_for_copy(&self, replicated_table_schema: &ReplicatedTableSchema, async_result: DropTableForCopyResult<()>) -> impl Future<Output = EtlResult<()>> + Send;
     fn write_table_rows(&self, replicated_table_schema: &ReplicatedTableSchema, rows: Vec<TableRow>, async_result: WriteTableRowsResult<()>) -> impl Future<Output = EtlResult<()>> + Send;
     fn write_events(&self, events: Vec<Event>, async_result: WriteEventsResult<()>) -> impl Future<Output = EtlResult<()>> + Send;
@@ -91,6 +92,8 @@ pub trait Destination {
 | Method | When called | Purpose |
 |--------|-------------|---------|
 | `name()` | On initialization | Identify the destination |
+| `shutdown()` | After ETL stops submitting work | Clean up destination resources, drain writers, or stop background tasks |
+| `startup()` | After store caches are loaded and before workers start | Reconcile destination state after process restarts |
 | `drop_table_for_copy()` | Before restarting a table copy when previous destination state exists | Drop the existing destination object and destination-private replay state using the previously stored replicated schema |
 | `write_table_rows()` | During initial copy | Receive bulk rows for the current replicated schema |
 | `write_events()` | During catch-up and continuous replication | Receive streaming changes |
